chore: adds tests, clippy, formatting, and auditing to CI
changes the following to appease the new CI rules:
  * ignores unused dep warning for openssl-sys.
  * runs cargo +nightly fmt on source.
  * fixes clippy issues.
  * removes modules from V1 that weren't being used. These modules were behind v1
    feature flags that had been removed as well.
  * adds deny.toml from c2pa-rs.

+++ src/test_stream.rs
@@ -57,8 +57,8 @@ impl Stream for TestStream {
             //stream.seek(SeekFrom::Start(pos as u64)).map_err(|e| StreamError::Io{ reason: e.to_string()})?;
             let whence = match mode {
                 SeekMode::Start => SeekFrom::Start(pos as u64),
-                SeekMode::End => SeekFrom::End(pos as i64),
-                SeekMode::Current => SeekFrom::Current(pos as i64),
+                SeekMode::End => SeekFrom::End(pos),
+                SeekMode::Current => SeekFrom::Current(pos),
             };
             stream.seek(whence).map_err(|e| Error::Io {
                 reason: e.to_string(),
